Project_Expense_Tracking_System/requirements.txt
streamlit==1.45.1
fastapi==0.115.10
uvicorn==0.34.0
pydantic==2.10.6
mysql-connector-python==9.3.0
numpy==2.2.1
pandas==2.2.3
requests==2.32.3
pytest==8.3.5


